fix bug with conv wstring, string and fix warning (correcting type for idx of children of node)

+++ nickname.h
@@ -165,7 +165,7 @@ class RadixTrie {
 		wstring endMark = node->isEnd ? L"$" : L"";
 		out << (isOutQuotes ? L"\"" : L"") << node->label << (isOutQuotes ? L"\"" : L"") << endMark << L'\n';
 		node->forEach([this, &out, &parentLines](
-			std::unique_ptr<Node>& node, [[maybe_unused]] int idx, bool isLast) {
+			std::unique_ptr<Node>& node, [[maybe_unused]] uint8_t idx, bool isLast) {
 			_printTree(out, node, false, isLast, parentLines);
 			});
 		if (!topLevel) {
@@ -203,7 +203,7 @@ class RadixTrie {
 				out << upFirstChar(curLabel) << " " << upFirstChar(path) << endl;
 			}
 		}
-		node->forEach([this, &curLabel, &out](std::unique_ptr<Node>& node, int idx, [[maybe_unused]] bool isLast) {
+		node->forEach([this, &curLabel, &out](std::unique_ptr<Node>& node, uint8_t idx, [[maybe_unused]] bool isLast) {
 			wchar_t ch = Node::getChar(idx);
 			wstring curPath = curLabel + wstring{ ch };
 			_print(out, node, curLabel, curPath);
